+++ src/pinfm/pinfm.cpp
@@ -140,7 +140,7 @@ namespace pinfm {
         AudioPipeline audio{*source,
                             {
                                 .loop               = params.loop,
-                                .targetSampleRate   = static_cast<int>(TARGET_SAMPLE_RATE),
+                                .targetSampleRate   = TARGET_SAMPLE_RATE,
                                 .targetOutputFrames = TARGET_OUTPUT_FRAMES,
                                 .tapsPerPhase       = RESAMPLER_TAPS_PER_PHASE,
                                 .maxCutoffHz        = RESAMPLER_LPF_CUTOFF,
@@ -153,7 +153,8 @@ namespace pinfm {
                   << peakDeviation << " Hz), loop=" << (params.loop ? "yes" : "no") << std::endl;
 
         NfmProcessor nfm{static_cast<float>(TARGET_SAMPLE_RATE), peakDeviation};
-        ngfmdmasync dma{params.transmissionFrequency, TARGET_SAMPLE_RATE, DMA_BIT_DEPTH, DMA_FIFO_SIZE};
+        ngfmdmasync dma{params.transmissionFrequency, static_cast<uint32_t>(TARGET_SAMPLE_RATE), DMA_BIT_DEPTH,
+                        DMA_FIFO_SIZE};
 
         // AudioPipeline owns source-rate input buffers, downmixing, loop-aware
         // EOF handling, and source -> 48 kHz rate conversion. outputBuf is reused
